feat(db): added coords for all the gathering nodes
