+++ apps/client/src/assets/data/monsters.json
@@ -21011,6 +21011,16 @@
         "x": 13,
         "y": 24.2,
         "z": 1.2
+      },
+      {
+        "map": 17,
+        "zoneid": 32,
+        "level": 0,
+        "hp": 887,
+        "fate": 0,
+        "x": 21.5,
+        "y": 36.3,
+        "z": 31782757990.3
       }
     ]
   },
@@ -53126,6 +53136,16 @@
         "x": 29.3,
         "y": 32,
         "z": 0.2
+      },
+      {
+        "map": 17,
+        "zoneid": 32,
+        "level": 0,
+        "hp": 887,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": -0.1
       }
     ]
   },
@@ -96346,6 +96366,16 @@
         "x": 21.1,
         "y": 25.9,
         "z": 0.7
+      },
+      {
+        "map": 369,
+        "zoneid": 2408,
+        "level": 0,
+        "hp": 29662,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
       }
     ]
   },
@@ -150591,6 +150621,146 @@
         "x": 37.6,
         "y": 23.8,
         "z": 1.1
+      }
+    ]
+  },
+  "10261": {
+    "baseid": 13498,
+    "positions": [
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 27054,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
+      }
+    ]
+  },
+  "10672": {
+    "baseid": 13392,
+    "positions": [
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 55066,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
+      }
+    ]
+  },
+  "10681": {
+    "baseid": 13402,
+    "positions": [
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 55066,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 675.8
+      },
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 55066,
+        "fate": 0,
+        "x": 21.5,
+        "y": 19.6,
+        "z": -16.2
+      },
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 55066,
+        "fate": 0,
+        "x": 21.5,
+        "y": -18.6,
+        "z": 0
+      }
+    ]
+  },
+  "10697": {
+    "baseid": 13524,
+    "positions": [
+      {
+        "map": 696,
+        "zoneid": 3709,
+        "level": 0,
+        "hp": 55066,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": -0.1
+      }
+    ]
+  },
+  "10698": {
+    "baseid": 13526,
+    "positions": [
+      {
+        "map": 696,
+        "zoneid": 3709,
+        "level": 0,
+        "hp": 57550,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
+      }
+    ]
+  },
+  "10704": {
+    "baseid": 13532,
+    "positions": [
+      {
+        "map": 696,
+        "zoneid": 3709,
+        "level": 0,
+        "hp": 57550,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
+      }
+    ]
+  },
+  "10713": {
+    "baseid": 13541,
+    "positions": [
+      {
+        "map": 696,
+        "zoneid": 3709,
+        "level": 0,
+        "hp": 62140,
+        "fate": 0,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
+      }
+    ]
+  },
+  "11081": {
+    "baseid": 14320,
+    "positions": [
+      {
+        "map": 695,
+        "zoneid": 3708,
+        "level": 0,
+        "hp": 3469172,
+        "fate": 1749,
+        "x": 21.5,
+        "y": 21.5,
+        "z": 0
       }
     ]
   },
